requirements.txt
streamlit
PyMuPDF
ebooklib
sentence-transformers
faiss-cpu
python-dotenv
requests
ebooklib
docx
python-docx